high level readme pass for project sanity
Added a protocol for discovering fully anonymous peers using beacons & lazy searchlights for the local subnet. It utilizes LLMs to negotiate an agreed upon protocol for future comms. I do not think that this protocol, or anything like it exists anywhere else.

So I've named it S.A.D.L. - Swarm Aware Discovery Logic. And the helping functions are SADL sniffers. We like to have fun here.

The key part of this logic is that ANY swarm, will willingly and openly communicate to ANY OTHER swarm discovered in the wild. And it will immediately be part of our swarm - low tust initilizations, but a path to trusted status.